requirements.txt
autogen-agentchat
autogen-ext[openai]
python-dotenv
requests
streamlit
python-pptx
python-docx
fpdf2
